Fixed tests and more small changes: - Streamlined logging and removed loggerBlock. We should switch to   os_log if available (iOS 10) and perhaps reconsider providing   a backwards compatible solution unless it's really compelling. - Re-indented TrustKit.h - Switched TrustKit to use GCD block callbacks for validation result   instead of notifications. It's a lot more performant to not dispatch   main thread notifications for every network requests unless they are   actually needed. It's also cleaner to not support loose coupling – app   can do whatever they want in the callback, including to send NSNotification.

+++ TrustKit/Pinning/TSKSPKIHashCache.m
@@ -101,7 +101,7 @@ static unsigned int asn1HeaderSizes[4] = { sizeof(rsa2048Asn1Header), sizeof(rsa
         // Initialize our cache of SPKI hashes
         // First try to load a cached version from the filesystem
         _subjectPublicKeyInfoHashesCache = [self getSpkiCacheFromFileSystem];
-        TSKLog(@"Loaded %d SPKI cache entries from the filesystem", _subjectPublicKeyInfoHashesCache.count);
+        TSKLog(@"Loaded %lu SPKI cache entries from the filesystem", (unsigned long)_subjectPublicKeyInfoHashesCache.count);
         
         if (_subjectPublicKeyInfoHashesCache == nil)
         {
